refactor: rename component 'History' -> 'HistoryData'
